# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chiptaol API Documentation",
        "description": "Open API Documentation for https://chiptaol.uz",
        "contact": {
            "email": "bakhadyrovf@gmail.com"
        },
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "https://chiptaol.uz"
        }
    ],
    "paths": {
...
        "/api/dashboard/movies/search/{title}": {
            "get": {
                "tags": [
                    "Dashboard Seances"
                ],
                "summary": "Get similar movies with provided title",
                "operationId": "76353da88cd0b413eb3f44eba7d0af0e",
                "parameters": [
                    {
                        "name": "title",
                        "in": "path",
                        "required": true,
                        "example": "Мстители"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success (OK) [Response](https://api.chiptaol.uz/api/example-responses/dashboard-movies-search)"
                    },
                    "400": {
                        "description": "Failure (Bad Request)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "default": "TMDB вернул неуспешный ответ, повторите еще раз."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/dashboard/cinemas/{cinema-id}/halls/{hall-id}/seats": {
            "get": {
                "tags": [
                    "Dashboard Halls"
                ],
                "summary": "Get a list of seats for specific hall",
                "operationId": "4ebf16a1346f62dc592ef008d77dfa80",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/cinema-id-path"
                    },
                    {
                        "$ref": "#/components/parameters/hall-id-path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success (OK) [Response](https://api.chiptaol.uz/api/example-responses/dashboard-seats)"
                    }
                }
            },
            "post": {
                "tags": [
                    "Dashboard Halls"
                ],
                "summary": "Add new seats for specific hall",
                "operationId": "a6abf2e010993879ff47a814a891e33b",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/cinema-id-path"
                    },
                    {
                        "$ref": "#/components/parameters/hall-id-path"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SeatStoreFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success (Created)"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Dashboard Halls"
                ],
                "summary": "Delete all seats for specific hall",
                "operationId": "f024f36dbe300f113ecccf938bdd25c1",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/cinema-id-path"
                    },
                    {
                        "$ref": "#/components/parameters/hall-id-path"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success (No Content)"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "SeatStoreFormRequest": {
                "properties": {
                    "seats": {
                        "type": "array",
                        "items": {
                            "required": [
                                "row",
                                "place",
                                "x",
                                "y"
                            ],
                            "properties": {
                                "is_vip": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "row": {
                                    "type": "integer",
                                    "example": 10
                                },
                                "place": {
                                    "type": "integer",
                                    "example": 30
                                },
                                "x": {
                                    "type": "integer",
                                    "example": 335
                                },
                                "y": {
                                    "type": "integer",
                                    "example": 90
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
            "cinema-id-path": {
                "name": "cinema-id",
                "in": "path",
                "description": "Cinema ID",
                "required": true
            },
            "hall-id-path": {
                "name": "hall-id",
                "in": "path",
                "description": "Hall ID",
                "required": true
            }
        }
    }
}